fix file import path after the restructure
