+++ src/components/newsFeed/NewsFeed.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Box } from "@mui/material";
-import Create from "../featuers/posts/Create";
-import PostsLists from "../featuers/posts/PostsList";
+import Create from "../../featuers/posts/Create";
+import PostsLists from "../../featuers/posts/PostsList";
 
 const NewsFeed = () => {
   const [currentId, setCurrentId] = useState(0);
